+++ app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/unlock/UnlockScreen.kt
@@ -42,8 +42,8 @@ import net.invictusmanagement.invictuskiosk.presentation.navigation.HomeScreen
 @Composable
 fun UnlockScreen(
     modifier: Modifier = Modifier,
-    unitId: Int,
-    mapId: Int,
+    unitId: Long,
+    mapId: Long,
     toPackageCenter: Boolean,
     navController: NavController,
     mainViewModel: MainViewModel = hiltViewModel()
@@ -54,10 +54,10 @@ fun UnlockScreen(
     val kioskName by mainViewModel.kioskName.collectAsStateWithLifecycle()
 
     LaunchedEffect(true) {
-        if (unitId != 0 && mapId != 0) {
+        if (unitId != 0L && mapId != 0L) {
             mainViewModel.fetchMapImage(
-                unitId = unitId.toLong(),
-                unitMapId = mapId.toLong(),
+                unitId = unitId,
+                unitMapId = mapId,
                 toPackageCenter = toPackageCenter
             )
         }
